requirements.txt
jupyter == 1.0.0
jupyter_contrib_nbextensions == 0.3.3
click == 6.7
pyyaml == 3.12
pytorch-sconce == 0.5.0
